Got cards working per route
