--- app/src/Directory/ClassList.css
@@ -1,7 +1,0 @@
-.card {
-    border: 1px solid #ccc;
-    top: 15%;
-    left: 20%;
-    right: 5%;
-  }
-
