+++ app/src/Homepage/Homepage.css
@@ -55,3 +55,17 @@
   height: 200vh;
   text-align: center;
 }
+
+.card {
+  border: 1px solid #ccc;
+  padding: 10px;
+  margin: 10px;
+  background-color: #fff;
+}
+
+.class-list {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+}
+
